refactor(boot): explicit config/client init and startup ping for DB/Redis
