+++ resources/application.properties
@@ -9,7 +9,6 @@ app.datasource.fetch-size=100
 app.datasource.max-rows=1000
 app.datasource.query-timeout=10
 app.datasource.pool.size=10
-app.datasource.pool.acquire-timeout-ms=3000
 # --- Cache (Redis)
 app.redis.host=localhost
 app.redis.port=6379
