+++ src/ru/andrewb/charm/back/bootstrap/SetupCheck.java
@@ -4,8 +4,11 @@ import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
 import jakarta.servlet.annotation.WebListener;
 import ru.andrewb.charm.back.dao.ProfileDao;
+import ru.andrewb.charm.back.dao.ProfileLikeDao;
 import ru.andrewb.charm.back.service.ContentService;
-import ru.andrewb.charm.back.utils.Config;
+import ru.andrewb.charm.back.config.Config;
+import ru.andrewb.charm.back.infra.db.ConnectionManager;
+import ru.andrewb.charm.back.infra.cache.RedisManager;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -15,19 +18,29 @@ public class SetupCheck implements ServletContextListener {
 
     @Override
     public void contextInitialized(ServletContextEvent sce) {
+        // 0) init config first
+        Config.init();
+
         // 1) check required configuration keys
         assertRequired(
                 "app.datasource.url",
                 "app.datasource.username",
                 "app.datasource.password",
-                "app.content.base-path"
+                "app.content.base-path",
+                "app.redis.host",
+                "app.redis.port"
         );
 
-        // 2) force singletons initialization
-        ContentService.getInstance();
-        ProfileDao dao = ProfileDao.getInstance();
+        // 2) init pools/clients explicitly
+        ConnectionManager.initOrThrow();
+        RedisManager.initOrThrow();
 
-        // 3) check if content base path exists (or create)
+        // 3) init singletons
+        ContentService.getInstance();
+        ProfileDao.getInstance();
+        ProfileLikeDao.getInstance();
+
+        // 4) check if content base path exists (or create)
         String basePathStr = Config.required("app.content.base-path");
         Path basePath = Path.of(basePathStr).toAbsolutePath().normalize();
         try {
@@ -36,8 +49,9 @@ public class SetupCheck implements ServletContextListener {
             throw new IllegalStateException("Content base path is invalid or not writable: " + basePath, e);
         }
 
-        // 4) ping DB
-        dao.ping();
+        // 5) ping DB and Redis
+        ConnectionManager.pingOrThrow();
+        RedisManager.pingOrThrow();
 
         sce.getServletContext().log(
                 "StartupCheck OK. Active profile=" + Config.getOrDefault("app.profile.active", "<none>")
