+++ src/ru/andrewb/charm/back/infra/cache/RedisManager.java
@@ -1,10 +1,11 @@
-package ru.andrewb.charm.back.config;
+package ru.andrewb.charm.back.infra.cache;
 
 import lombok.experimental.UtilityClass;
 import lombok.extern.slf4j.Slf4j;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisPool;
 import redis.clients.jedis.JedisPoolConfig;
+import ru.andrewb.charm.back.config.Config;
 
 @Slf4j
 @UtilityClass
